requirements.txt
pyTelegramBotAPI
python-dotenv
transformers
torch
